feat(shader): refactor interface & more specific linting info for undefined variable
